# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ecommerse",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "variables": {}
    }
  ],
  "paths": {
...
    "/product/get/658d98006e004542f0d84089": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get One Data",
        "operationId": "getOneData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/user/get/658d99c26e004542f0d840ac": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "get one data",
        "operationId": "getonedata",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}